+++ tests/data/index.js
@@ -13,7 +13,8 @@ import $M from './form.m.js';
 import $N from './form.n.js';
 import $O from './form.o.js';
 import $P from './form.p.js';
+import $Q from './form.q.js';
 
 export const up = str => str.toUpperCase();
 
-export default { $A, $B, $C, $D, $E, $F, $G, $H, $I, $L, $M, $N, $O, $P };
+export default { $A, $B, $C, $D, $E, $F, $G, $H, $I, $L, $M, $N, $O, $P, $Q };